modularize lexical-html source
