+++ packages/lexical-html/src/__tests__/unit/DOMExtension.test.ts
@@ -7,13 +7,7 @@
  */
 
 import {buildEditorFromExtensions} from '@lexical/extension';
-import {
-  $generateDOMFromRoot,
-  $getDOMContextValue,
-  DOMContextRoot,
-  DOMExtension,
-  domOverride,
-} from '@lexical/html';
+import {$generateDOMFromRoot} from '@lexical/html';
 import {
   $createParagraphNode,
   $createTextNode,
@@ -29,6 +23,10 @@ import {
 } from 'lexical';
 import {expectHtmlToBeEqual, html} from 'lexical/src/__tests__/utils';
 import {describe, expect, test} from 'vitest';
+
+import {$getDOMContextValue, DOMContextRoot} from '../../ContextRecord';
+import {DOMExtension} from '../../DOMExtension';
+import {domOverride} from '../../domOverride';
 
 const idState = createState('id', {
   parse: (v) => (typeof v === 'string' ? v : null),
