+++ packages/lexical-html/src/constants.ts
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+export const DOMExtensionName = '@lexical/html/DOM';
+export const DOMImportExtensionName = '@lexical/html/DOMImport';
+export const IGNORE_TAGS = new Set(['STYLE', 'SCRIPT']);
